flights updated

+++ src/containers/regions/Regions.jsx
@@ -2,11 +2,10 @@ import React, {useEffect, useState} from 'react';
 import regions from "../../api/projectApi/regions";
 import moment from "moment";
 import {GrTrash} from "react-icons/gr";
-import UserModal from "../../components/modal/UserModal";
 import {userRole} from "../../constants/userRole";
 import RegionModal from "../../components/modal/RegionModal";
 
-function Regions(props) {
+function Regions() {
     const [regions_list, setRegions] = useState([]);
 
     useEffect(() => {
@@ -14,36 +13,33 @@ function Regions(props) {
     }, []);
 
     const handleDelete = (id) => {
-        regions.deleteOne(id)
-            .then(res => regions_list.filter(a => a.id !== id))
-            .catch(res => regions_list.filter(a => a.id !== id))
-        regions.getAllRegions().then(res => {
-            setRegions(res.data.result[0]);
-            console.log(res.data.result[0])
-        })
-    }
+        const confirmation = window.confirm("Are you sure to delete this region?");
+        if (confirmation) {
+            regions.deleteOne(id)
+                .then(() => regions_list.filter(a => a.id !== id))
+                .catch(() => regions_list.filter(a => a.id !== id))
+            regions.getAllRegions().then(res => {
+                setRegions(res.data.result[0]);
+                console.log(res.data.result[0])
+            })
+        }
+    };
 
     const [userModal, setUserModal] = useState({
-        typeModal: "create",
-        openModal: false,
-        values: {}
+        typeModal: "create", openModal: false, values: {}
     })
 
 
     const handleClose = () => {
         setUserModal({
-            typeModal: "create",
-            openModal: false,
-            values: {}
+            typeModal: "create", openModal: false, values: {}
         })
     }
 
     const handlePressItemEdit = (type, item) => {
         console.log(item)
         setUserModal({
-            typeModal: type,
-            openModal: true,
-            values: item
+            typeModal: type, openModal: true, values: item
         })
     }
 
@@ -55,23 +51,27 @@ function Regions(props) {
                     регион
                 </button>
             </div>
-            <table className={"w-full border"}>
+            <table className={"w-full border border-red-200"}>
                 <thead>
-                <tr className={"border"}>
-                    <th className={"p-2 border"}>Name</th>
-                    <th className={"p-2 border"}>Created At</th>
-                    <th className={"p-2 border"}>Actions</th>
+                <tr className={"border border-red-200"}>
+                    <th className={"p-2 border border-red-200"}>Name</th>
+                    <th className={"p-2 border border-red-200"}>Created At</th>
+                    <th className={"p-2 border border-red-200"}>Actions</th>
                 </tr>
                 </thead>
                 <tbody>
-                {regions_list.map(a => <tr key={a.id} className={"border"}>
-                    <td className={"border p-2"}>{a.name}</td>
-                    <td className={"border p-2"}>{moment(a.createdAt).format("MMMM DD YYYY HH:MM")}</td>
-                    <td className={"border p-2"}>
-                        <button className={"p-2 border rounded hover:bg-gray-200 transition"}
-                                onClick={() => handleDelete(a.id)}><GrTrash fontSize="1.5em"/></button>
-                    </td>
-                </tr>)}
+                {regions_list.length === 0 ? <tr className={"border border-red-200"}>
+                    <td colSpan={3} className={"border border-red-200 p-2 text-center"}>No Data</td>
+                </tr> : regions_list.map(a => {
+                    return (<tr key={a.id} className={"border border-red-200"}>
+                        <td className={"border border-red-200 p-2"}>{a.name}</td>
+                        <td className={"border border-red-200 p-2"}>{moment(a.createdAt).format("MMMM DD YYYY HH:MM")}</td>
+                        <td className={"border border-red-200 p-2"}>
+                            <button className={"p-2 border border-red-200 rounded hover:bg-gray-200 transition"}
+                                    onClick={() => handleDelete(a.id)}><GrTrash fontSize="1.5em"/></button>
+                        </td>
+                    </tr>)
+                })}
                 </tbody>
             </table>
         </div>
@@ -82,11 +82,6 @@ function Regions(props) {
             values={userModal.values}
             setRegions={setRegions}
         />
-        {/*<input type="text" placeholder={"Region Name"}/>*/}
-        {/*<input type="text" placeholder={"Region Name"}/>*/}
-        {/*<input type="text" placeholder={"Region Name"}/>*/}
-        {/*<input type="text" placeholder={"Region Name"}/>*/}
-        {/*<input type="text" placeholder={"Region Name"}/>*/}
     </div>);
 }
